move initializer and adjust visibility accordingly
